+++ src/Controller/FO/TempsController.php
@@ -4,8 +4,9 @@ namespace App\Controller\FO;
 
 use App\Exception\MonthOutOfRangeException;
 use App\Form\TempsPassesType;
+use App\Service\CraService;
 use App\Service\DateMonthService;
-use App\Service\TempsPasseService;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
@@ -23,7 +24,8 @@ class TempsController extends AbstractController
         Request $request,
         string $year = null,
         string $month = null,
-        TempsPasseService $tempsPasseService,
+        CraService $craService,
+        EntityManagerInterface $em,
         DateMonthService $dateMonthService
     ) {
         if ($year !== null && $month === null) {
@@ -40,18 +42,15 @@ class TempsController extends AbstractController
             throw $this->createNotFoundException('Impossible de saisir les temps passés dans le futur.');
         }
 
-        $listeTempsPasses = $tempsPasseService->loadTempsPasses($this->getUser(), $mois);
-        $form = $this->createForm(TempsPassesType::class, $listeTempsPasses);
+        $cra = $craService->loadCraForUser($this->getUser(), $mois);
+        $form = $this->createForm(TempsPassesType::class, $cra);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
+            $cra->setTempsPassesModifiedAt(new \DateTime());
 
-            foreach ($listeTempsPasses->getTempsPasses() as $tempsPasse) {
-                $em->persist($tempsPasse);
-            }
-
+            $em->persist($cra);
             $em->flush();
 
             $href = $this->generateUrl('absences_', [
@@ -76,7 +75,7 @@ class TempsController extends AbstractController
             'form' => $form->createView(),
             'next' => $dateMonthService->getNextMonth($mois),
             'prev' => $dateMonthService->getPrevMonth($mois),
-            'hasTempsPasses' => $listeTempsPasses->hasTempsPasses(),
+            'cra' => $cra,
         ]);
     }
 
@@ -88,6 +87,7 @@ class TempsController extends AbstractController
      * )
      */
     public function saisieAbsences(
+        CraService $craService,
         string $year = null,
         string $month = null,
         DateMonthService $dateMonthService
@@ -103,11 +103,11 @@ class TempsController extends AbstractController
         }
 
         return $this->render('temps/absences.html.twig', [
-            'date' => $date,
             'next' => $dateMonthService->getNextMonth($date),
             'prev' => $dateMonthService->getPrevMonth($date),
             'year' => $date->format('Y'),
             'month' => $date->format('m'),
+            'cra' => $craService->loadCraForUser($this->getUser(), $date),
         ]);
     }
 }